Extract llm-related code from mng_claude_mind into new mng_llm plugin
Creates a new mng_llm plugin that provides:
- LlmAgent class (inherits from BaseAgent) for running the llm CLI tool
- LlmAgentConfig with llm-specific configuration
- All llm-related provisioning: toolchain installation, conversation
  management, supporting services (chat.sh, ttyd_chat.sh), llm tools
  (context_tool.py, extra_context_tool.py)
- CLI commands: llmconversations, llmweb, llmdb
- Settings: ChatSettings, ContextSettings, ExtraContextSettings,
  ProvisioningSettings
- Resources: conversation_watcher.py, web_server.py, conversation_db.py

Also refactors shared infrastructure:
- Moves watcher_common.py from mng_claude_mind to mng_recursive so
  both mng_llm and mng_claude_mind can depend on it
- Moves ttyd_agent.sh from mng_claude_mind to mng_ttyd with a new
  on_after_provisioning hook
- Removes create_event_log_directories (scripts create dirs directly)

Dependency chain: mng_claude_mind -> mng_llm -> mng_recursive
