+++ libs/mng_claude_mind/imbue/mng_claude_mind/data_types.py
@@ -10,17 +10,14 @@ from imbue.imbue_common.event_envelope import EventSource
 from imbue.imbue_common.frozen_model import FrozenModel
 from imbue.imbue_common.primitives import NonEmptyStr
 from imbue.imbue_common.primitives import NonNegativeInt
-from imbue.imbue_common.primitives import PositiveFloat
 from imbue.imbue_common.primitives import PositiveInt
-from imbue.mng_claude_mind.resources.watcher_common import DEFAULT_CEL_FILTER
+from imbue.mng_llm.data_types import ChatModel
+from imbue.mng_llm.data_types import LlmSettings
+from imbue.mng_recursive.watcher_common import DEFAULT_CEL_FILTER
 
 
 class ConversationId(NonEmptyStr):
     """Unique identifier for a conversation thread (matches llm's conversation_id format)."""
-
-
-class ChatModel(NonEmptyStr):
-    """Model name used for chat conversations (e.g. 'claude-sonnet-4-6')."""
 
 
 class MessageRole(NonEmptyStr):
@@ -42,12 +39,7 @@ SOURCE_COMMON_TRANSCRIPT: Final[EventSource] = EventSource("common_transcript")
 
 
 class MessageEvent(EventEnvelope):
-    """An event in events/messages/events.jsonl recording a conversation message.
-
-    Each event represents a single user or assistant message. All messages
-    across all conversations go into the same file, with conversation_id
-    identifying which conversation the message belongs to.
-    """
+    """An event in events/messages/events.jsonl recording a conversation message."""
 
     conversation_id: ConversationId
     role: MessageRole
@@ -57,25 +49,12 @@ class MessageEvent(EventEnvelope):
 class MindEvent(EventEnvelope):
     """A generic event with a data payload, used for sources like scheduled,
     mng/agents, stop, and monitor.
-
-    The data field carries event-type-specific payload.
     """
 
     data: dict[str, Any] = {}
 
 
 # -- Common transcript types --
-# These define the agent-agnostic common message format written to
-# events/common_transcript/events.jsonl. Inspired by oh-my-pi's session
-# entry format, but adapted to our EventEnvelope conventions.
-#
-# The common format focuses on semantically important messages (user input,
-# assistant output, tool calls/results) and drops noise like progress events,
-# file-history snapshots, and system bookkeeping.
-#
-# NOTE: These types document the schema produced by transcript_watcher.py.
-# The watcher runs as a standalone script on the host and produces matching
-# JSON directly (it cannot import these classes).
 
 
 class CommonToolCallSummary(FrozenModel):
@@ -100,11 +79,7 @@ class CommonTokenUsage(FrozenModel):
 
 
 class CommonUserMessageEvent(EventEnvelope):
-    """A user message in the common transcript.
-
-    Emitted for each Claude ``type: "user"`` event that contains actual
-    user-authored text (not tool results).
-    """
+    """A user message in the common transcript."""
 
     role: MessageRole = Field(default=MessageRole("user"), description="Always 'user'.")
     content: str = Field(description="The user's message text.")
@@ -114,11 +89,7 @@ class CommonUserMessageEvent(EventEnvelope):
 
 
 class CommonAssistantMessageEvent(EventEnvelope):
-    """An assistant response in the common transcript.
-
-    Emitted for each Claude ``type: "assistant"`` event. Contains the
-    concatenated text output, any tool calls, model info, and token usage.
-    """
+    """An assistant response in the common transcript."""
 
     role: MessageRole = Field(default=MessageRole("assistant"), description="Always 'assistant'.")
     model: ChatModel = Field(description="Model that generated this response (e.g. 'claude-opus-4.6').")
@@ -132,11 +103,7 @@ class CommonAssistantMessageEvent(EventEnvelope):
 
 
 class CommonToolResultEvent(EventEnvelope):
-    """A tool result in the common transcript.
-
-    Emitted for each tool result block found inside Claude ``type: "user"``
-    events. These are the return values from tool invocations.
-    """
+    """A tool result in the common transcript."""
 
     tool_call_id: NonEmptyStr = Field(description="ID of the tool call this result corresponds to.")
     tool_name: NonEmptyStr = Field(description="Name of the tool (resolved from the preceding assistant message).")
@@ -150,70 +117,6 @@ class CommonToolResultEvent(EventEnvelope):
 # -- Settings types --
 # These model the structure of minds.toml.
 # Each section corresponds to a TOML table.
-
-
-class ContextSettings(FrozenModel):
-    """Settings for the [chat.context] TOML section (used by context_tool.py)."""
-
-    max_transcript_line_count: PositiveInt = Field(
-        default=PositiveInt(10),
-        description="Maximum number of inner monologue lines to include in context.",
-    )
-    max_messages_line_count: PositiveInt = Field(
-        default=PositiveInt(20),
-        description="Maximum number of recent message lines to include in context.",
-    )
-    max_messages_per_conversation: PositiveInt = Field(
-        default=PositiveInt(3),
-        description="Maximum messages to show per conversation in context.",
-    )
-    max_trigger_line_count: PositiveInt = Field(
-        default=PositiveInt(5),
-        description="Maximum trigger event lines per source in context.",
-    )
-    max_content_length: PositiveInt = Field(
-        default=PositiveInt(200),
-        description="Maximum character length for truncated event content in context_tool.",
-    )
-
-
-class ExtraContextSettings(FrozenModel):
-    """Settings for the [chat.extra_context] TOML section (used by extra_context_tool.py)."""
-
-    max_content_length: PositiveInt = Field(
-        default=PositiveInt(300),
-        description="Maximum character length for truncated event content in extra_context_tool.",
-    )
-    transcript_line_count: PositiveInt = Field(
-        default=PositiveInt(50),
-        description="Number of inner monologue lines to show in extended history.",
-    )
-    mng_list_hard_timeout_seconds: PositiveFloat = Field(
-        default=PositiveFloat(120.0),
-        description="Hard timeout for mng list command (seconds).",
-    )
-    mng_list_warn_threshold_seconds: PositiveFloat = Field(
-        default=PositiveFloat(15.0),
-        description="Warning threshold for mng list command (seconds).",
-    )
-
-
-class ChatSettings(FrozenModel):
-    """Settings for the [chat] TOML section."""
-
-    model: ChatModel | None = Field(
-        default=None,
-        description="Default model for new conversation threads. "
-        "When None, chat.sh falls back to the hardcoded default (claude-opus-4.6).",
-    )
-    context: ContextSettings = Field(
-        default_factory=ContextSettings,
-        description="Context tool settings ([chat.context] section).",
-    )
-    extra_context: ExtraContextSettings = Field(
-        default_factory=ExtraContextSettings,
-        description="Extra context tool settings ([chat.extra_context] section).",
-    )
 
 
 class WatcherSettings(FrozenModel):
@@ -255,40 +158,12 @@ class WatcherSettings(FrozenModel):
     )
 
 
-class ProvisioningSettings(FrozenModel):
-    """Settings for the [provisioning] TOML section."""
-
-    fs_hard_timeout_seconds: PositiveFloat = Field(
-        default=PositiveFloat(16.0),
-        description="Hard timeout for filesystem operations (seconds).",
-    )
-    fs_warn_threshold_seconds: PositiveFloat = Field(
-        default=PositiveFloat(4.0),
-        description="Warning threshold for filesystem operations (seconds).",
-    )
-    command_check_hard_timeout_seconds: PositiveFloat = Field(
-        default=PositiveFloat(30.0),
-        description="Hard timeout for command existence checks (seconds).",
-    )
-    command_check_warn_threshold_seconds: PositiveFloat = Field(
-        default=PositiveFloat(5.0),
-        description="Warning threshold for command existence checks (seconds).",
-    )
-    install_hard_timeout_seconds: PositiveFloat = Field(
-        default=PositiveFloat(300.0),
-        description="Hard timeout for package installations (seconds).",
-    )
-    install_warn_threshold_seconds: PositiveFloat = Field(
-        default=PositiveFloat(60.0),
-        description="Warning threshold for package installations (seconds).",
-    )
-
-
-class ClaudeMindSettings(FrozenModel):
+class ClaudeMindSettings(LlmSettings):
     """Top-level settings loaded from minds.toml.
 
-    All fields have defaults, so an empty or missing settings file
-    produces a valid settings object with the standard defaults.
+    Extends LlmSettings (chat, provisioning) with mind-specific sections
+    (agent_type, watchers). All fields have defaults, so an empty or missing
+    settings file produces a valid settings object with the standard defaults.
     """
 
     agent_type: str | None = Field(
@@ -296,15 +171,7 @@ class ClaudeMindSettings(FrozenModel):
         description="Agent type for this mind (e.g. 'elena-code', 'claude-mind'). "
         "Used by mind deploy to determine the agent type when --agent-type is not provided on the CLI.",
     )
-    chat: ChatSettings = Field(
-        default_factory=ChatSettings,
-        description="Chat-related settings ([chat] section).",
-    )
     watchers: WatcherSettings = Field(
         default_factory=WatcherSettings,
         description="Watcher settings ([watchers] section).",
-    )
-    provisioning: ProvisioningSettings = Field(
-        default_factory=ProvisioningSettings,
-        description="Provisioning timeout settings ([provisioning] section).",
     )
