+++ libs/mng_claude_mind/imbue/mng_claude_mind/provisioning.py
@@ -1,11 +1,16 @@
+"""Mind-specific provisioning functions for the claude-mind agent type.
+
+Provides provisioning for mind defaults (GLOBAL.md, role prompts, skills),
+symlinks, memory directory setup, and talking role constraint validation.
+
+LLM-related provisioning (toolchain installation, conversation management,
+supporting services) is provided by the mng_llm plugin.
+"""
+
 from __future__ import annotations
 
 import importlib.resources
-import json
 import shlex
-import time
-from datetime import datetime
-from datetime import timezone
 from pathlib import Path
 from typing import Any
 from typing import Final
@@ -13,31 +18,10 @@ from typing import Final
 from loguru import logger
 
 from imbue.imbue_common.logging import log_span
-from imbue.mng.interfaces.data_types import CommandResult
 from imbue.mng.interfaces.host import OnlineHostInterface
-from imbue.mng.providers.ssh_host_setup import load_resource_script
 from imbue.mng_claude_mind import resources as mind_resources
-from imbue.mng_claude_mind.data_types import ProvisioningSettings
-
-# Supporting service shell scripts to provision to $MNG_AGENT_STATE_DIR/commands/.
-# Python scripts (event_watcher, conversation_watcher, transcript_watcher,
-# web_server, conversation_db) are now registered as mng CLI commands and
-# do not need to be provisioned.
-_SERVICE_SCRIPT_FILES: Final[tuple[str, ...]] = ("chat.sh",)
-
-# Scripts provisioned to $MNG_AGENT_STATE_DIR/commands/ttyd/ for URL-arg dispatch.
-# Each <key>.sh is discovered by the ttyd port wrapper and registered as a virtual server.
-# Tuples of (resource filename, target filename under commands/ttyd/).
-_TTYD_DISPATCH_SCRIPTS: Final[tuple[tuple[str, str], ...]] = (
-    ("ttyd_agent.sh", "agent.sh"),
-    ("ttyd_chat.sh", "chat.sh"),
-)
-
-# Python tool files to provision to $MNG_AGENT_STATE_DIR/commands/llm_tools/
-_LLM_TOOL_FILES: Final[tuple[str, ...]] = (
-    "context_tool.py",
-    "extra_context_tool.py",
-)
+from imbue.mng_llm.data_types import ProvisioningSettings
+from imbue.mng_llm.provisioning import execute_with_timing
 
 # Default content files written to the work directory root if missing.
 # Tuples of (resource path under defaults/, target path relative to work dir).
@@ -65,28 +49,6 @@ _DEFAULT_SKILL_DIRS: Final[tuple[str, ...]] = (
 )
 
 
-def _execute_with_timing(
-    host: OnlineHostInterface,
-    cmd: str,
-    *,
-    hard_timeout: float,
-    warn_threshold: float,
-    label: str,
-) -> CommandResult:
-    """Execute a host command with two-threshold timeout monitoring.
-
-    Uses hard_timeout as the actual timeout. If the command takes longer
-    than warn_threshold, emits a warning so we can notice degradation
-    before it becomes an outright failure.
-    """
-    start = time.monotonic()
-    result = host.execute_command(cmd, timeout_seconds=hard_timeout)
-    elapsed = time.monotonic() - start
-    if elapsed > warn_threshold:
-        logger.warning("{} took {:.1f}s (expected <{:.0f}s): {}", label, elapsed, warn_threshold, cmd)
-    return result
-
-
 def load_mind_resource(filename: str) -> str:
     """Load a resource file from the mng_claude_mind resources package."""
     resource_files = importlib.resources.files(mind_resources)
@@ -101,7 +63,7 @@ def _write_default_if_missing(
     settings: ProvisioningSettings,
 ) -> None:
     """Write a default resource file to the host if the target doesn't already exist."""
-    check = _execute_with_timing(
+    check = execute_with_timing(
         host,
         f"test -f {shlex.quote(str(target_path))}",
         hard_timeout=settings.fs_hard_timeout_seconds,
@@ -112,7 +74,7 @@ def _write_default_if_missing(
         logger.debug("Default file already exists, skipping: {}", target_path)
         return
 
-    _execute_with_timing(
+    execute_with_timing(
         host,
         f"mkdir -p {shlex.quote(str(target_path.parent))}",
         hard_timeout=settings.fs_hard_timeout_seconds,
@@ -153,8 +115,7 @@ def validate_talking_role_constraints(
     talking_dir = work_dir / "talking"
     for name in _TALKING_FORBIDDEN:
         target = talking_dir / name
-        # Use -e so we catch both files and directories (including symlinks)
-        check = _execute_with_timing(
+        check = execute_with_timing(
             host,
             f"test -e {shlex.quote(str(target))}",
             hard_timeout=settings.fs_hard_timeout_seconds,
@@ -208,55 +169,6 @@ def provision_default_content(
         )
 
 
-def install_llm_toolchain(host: OnlineHostInterface, settings: ProvisioningSettings) -> None:
-    """Install llm, llm-anthropic, and llm-live-chat on the host.
-
-    Uses uv tool install for llm itself, then llm install for plugins.
-    Skips installation if llm is already available.
-    """
-    with log_span("Installing llm toolchain"):
-        # Check if llm is already installed
-        check_result = _execute_with_timing(
-            host,
-            "command -v llm",
-            hard_timeout=settings.command_check_hard_timeout_seconds,
-            warn_threshold=settings.command_check_warn_threshold_seconds,
-            label="llm check",
-        )
-        if check_result.success:
-            # llm is installed, just ensure plugins are present
-            _install_llm_plugins(host, settings)
-            return
-
-        # Install llm via uv tool
-        result = _execute_with_timing(
-            host,
-            "uv tool install llm",
-            hard_timeout=settings.install_hard_timeout_seconds,
-            warn_threshold=settings.install_warn_threshold_seconds,
-            label="llm install",
-        )
-        if not result.success:
-            raise RuntimeError(f"Failed to install llm: {result.stderr}")
-
-        _install_llm_plugins(host, settings)
-
-
-def _install_llm_plugins(host: OnlineHostInterface, settings: ProvisioningSettings) -> None:
-    """Install the required llm plugins for the models and features we use."""
-    for plugin_name in ("llm-anthropic", "llm-live-chat", "llm-matched-responses"):
-        with log_span("Installing llm plugin: {}", plugin_name):
-            result = _execute_with_timing(
-                host,
-                f"llm install {plugin_name}",
-                hard_timeout=settings.install_hard_timeout_seconds,
-                warn_threshold=settings.install_warn_threshold_seconds,
-                label=f"llm plugin install ({plugin_name})",
-            )
-            if not result.success:
-                raise RuntimeError(f"Failed to install {plugin_name}: {result.stderr}")
-
-
 def create_mind_symlinks(
     host: OnlineHostInterface,
     work_dir: Path,
@@ -268,12 +180,9 @@ def create_mind_symlinks(
     Claude Code runs from within the role directory (via ``cd $ROLE`` in
     assemble_command), so ``.claude/`` is found naturally. We only need:
 
-    - ``<work_dir>/CLAUDE.md`` -> ``<work_dir>/GLOBAL.md`` (found by Claude Code
-      walking up the directory tree from the role directory)
+    - ``<work_dir>/CLAUDE.md`` -> ``<work_dir>/GLOBAL.md``
     - ``<work_dir>/<active_role>/CLAUDE.local.md`` -> ``<work_dir>/<active_role>/PROMPT.md``
-      (role-specific prompt, discovered in the role directory)
     """
-    # CLAUDE.md -> GLOBAL.md (so Claude Code loads global instructions)
     _create_symlink_if_target_exists(
         host,
         link_path=work_dir / "CLAUDE.md",
@@ -281,7 +190,6 @@ def create_mind_symlinks(
         settings=settings,
     )
 
-    # <role>/CLAUDE.local.md -> <role>/PROMPT.md (role-specific prompt)
     _create_symlink_if_target_exists(
         host,
         link_path=work_dir / active_role / "CLAUDE.local.md",
@@ -297,7 +205,7 @@ def _create_symlink_if_target_exists(
     settings: ProvisioningSettings,
 ) -> None:
     """Create a symlink at link_path pointing to target_path, if target exists."""
-    check = _execute_with_timing(
+    check = execute_with_timing(
         host,
         f"test -f {shlex.quote(str(target_path))}",
         hard_timeout=settings.fs_hard_timeout_seconds,
@@ -307,8 +215,7 @@ def _create_symlink_if_target_exists(
     if not check.success:
         return
 
-    # Ensure parent directory exists
-    _execute_with_timing(
+    execute_with_timing(
         host,
         f"mkdir -p {shlex.quote(str(link_path.parent))}",
         hard_timeout=settings.fs_hard_timeout_seconds,
@@ -316,10 +223,9 @@ def _create_symlink_if_target_exists(
         label="mkdir",
     )
 
-    # Create symlink (force to overwrite existing)
     cmd = f"ln -sf {shlex.quote(str(target_path))} {shlex.quote(str(link_path))}"
     with log_span("Creating symlink: {} -> {}", link_path, target_path):
-        result = _execute_with_timing(
+        result = execute_with_timing(
             host,
             cmd,
             hard_timeout=settings.fs_hard_timeout_seconds,
@@ -330,407 +236,6 @@ def _create_symlink_if_target_exists(
             raise RuntimeError(f"Failed to create symlink {link_path} -> {target_path}: {result.stderr}")
 
 
-def provision_supporting_services(
-    host: OnlineHostInterface,
-    agent_state_dir: Path,
-    settings: ProvisioningSettings,
-) -> None:
-    """Write supporting service shell scripts to $MNG_AGENT_STATE_DIR/commands/.
-
-    Provisions:
-    - Shared logging library (mng_log.sh)
-    - Service scripts to commands/ (e.g. chat.sh)
-    - Ttyd dispatch scripts to commands/ttyd/ (e.g. agent.sh, chat.sh)
-
-    Python supporting services (event_watcher, conversation_watcher, transcript_watcher,
-    web_server, conversation_db) are registered as mng CLI commands and do not need
-    to be provisioned.
-    """
-    commands_dir = agent_state_dir / "commands"
-    ttyd_dir = commands_dir / "ttyd"
-    _execute_with_timing(
-        host,
-        f"mkdir -p {shlex.quote(str(ttyd_dir))}",
-        hard_timeout=settings.fs_hard_timeout_seconds,
-        warn_threshold=settings.fs_warn_threshold_seconds,
-        label="mkdir commands/ttyd",
-    )
-
-    # Provision the shared logging library (from mng core resources) first,
-    # since the supporting service scripts source it.
-    mng_log_content = load_resource_script("mng_log.sh")
-    mng_log_path = commands_dir / "mng_log.sh"
-    with log_span("Writing mng_log.sh to host"):
-        host.write_file(mng_log_path, mng_log_content.encode(), mode="0755")
-
-    for script_name in _SERVICE_SCRIPT_FILES:
-        script_content = load_mind_resource(script_name)
-        script_path = commands_dir / script_name
-        with log_span("Writing {} to host", script_name):
-            host.write_file(script_path, script_content.encode(), mode="0755")
-
-    for resource_name, target_name in _TTYD_DISPATCH_SCRIPTS:
-        script_content = load_mind_resource(resource_name)
-        script_path = ttyd_dir / target_name
-        with log_span("Writing ttyd/{} to host", target_name):
-            host.write_file(script_path, script_content.encode(), mode="0755")
-
-
-def provision_llm_tools(
-    host: OnlineHostInterface,
-    agent_state_dir: Path,
-    settings: ProvisioningSettings,
-) -> None:
-    """Write LLM tool Python files to $MNG_AGENT_STATE_DIR/commands/llm_tools/.
-
-    These files are passed to `llm live-chat` via `--functions` to give
-    conversation agents access to mind context.
-    """
-    tools_dir = agent_state_dir / "commands" / "llm_tools"
-    _execute_with_timing(
-        host,
-        f"mkdir -p {shlex.quote(str(tools_dir))}",
-        hard_timeout=settings.fs_hard_timeout_seconds,
-        warn_threshold=settings.fs_warn_threshold_seconds,
-        label="mkdir llm_tools",
-    )
-
-    for tool_file in _LLM_TOOL_FILES:
-        tool_content = load_mind_resource(tool_file)
-        tool_path = tools_dir / tool_file
-        with log_span("Writing {} to host", tool_file):
-            host.write_file(tool_path, tool_content.encode(), mode="0644")
-
-
-def create_event_log_directories(
-    host: OnlineHostInterface,
-    agent_state_dir: Path,
-    settings: ProvisioningSettings,
-) -> None:
-    """Create the event and log directory structure.
-
-    Creates directories for event sources (events/<source>/):
-    - events/messages/           conversation messages
-    - events/scheduled/          scheduled trigger events
-    - events/mng/agents/         agent state transitions
-    - events/stop/               agent stop events
-    - events/monitor/            (future) monitor agent events
-    - events/delivery_failures/  event delivery failure notifications
-    - events/common_transcript/  agent-agnostic transcript (written by transcript watcher)
-    - events/servers/            server registration records
-
-    Creates directories for log sources (logs/<source>/):
-    - logs/claude_transcript/    inner monologue (written by Claude background tasks, raw format)
-
-    Note: conversation metadata (tags, created_at) is stored in the
-    mind_conversations table in the llm sqlite database, not in
-    a separate event directory.
-    """
-    for source in (
-        "messages",
-        "scheduled",
-        "mng/agents",
-        "stop",
-        "monitor",
-        "delivery_failures",
-        "common_transcript",
-        "servers",
-    ):
-        source_dir = agent_state_dir / "events" / source
-        _execute_with_timing(
-            host,
-            f"mkdir -p {shlex.quote(str(source_dir))}",
-            hard_timeout=settings.fs_hard_timeout_seconds,
-            warn_threshold=settings.fs_warn_threshold_seconds,
-            label=f"mkdir events/{source}",
-        )
-
-    # Create log directories for raw/non-envelope data
-    for log_source in ("claude_transcript",):
-        log_dir = agent_state_dir / "logs" / log_source
-        _execute_with_timing(
-            host,
-            f"mkdir -p {shlex.quote(str(log_dir))}",
-            hard_timeout=settings.fs_hard_timeout_seconds,
-            warn_threshold=settings.fs_warn_threshold_seconds,
-            label=f"mkdir logs/{log_source}",
-        )
-
-
-def configure_llm_user_path(
-    host: OnlineHostInterface,
-    agent_state_dir: Path,
-    settings: ProvisioningSettings,
-) -> None:
-    """Create the per-agent llm data directory.
-
-    Creates ``<agent_state_dir>/llm_data/`` so that llm commands have a
-    unique database directory. The ``LLM_USER_PATH`` env var itself is
-    set by the host's ``_collect_agent_env_vars`` (in host.py), which
-    runs after ``agent.provision()`` and writes it to the agent env file
-    that gets sourced by all shell processes.
-    """
-    llm_data_dir = agent_state_dir / "llm_data"
-    _execute_with_timing(
-        host,
-        f"mkdir -p {shlex.quote(str(llm_data_dir))}",
-        hard_timeout=settings.fs_hard_timeout_seconds,
-        warn_threshold=settings.fs_warn_threshold_seconds,
-        label="mkdir llm_data",
-    )
-    logger.info("Created LLM data directory: {}", llm_data_dir)
-
-
-def _get_llm_db_path(agent_state_dir: Path) -> Path:
-    """Return the path to the llm database for the given agent."""
-    return agent_state_dir / "llm_data" / "logs.db"
-
-
-# SQL schema for the mind_conversations table that stores conversation
-# metadata (tags, created_at) alongside the llm tool's own tables.
-MIND_CONVERSATIONS_TABLE_SQL: Final[str] = (
-    "CREATE TABLE IF NOT EXISTS mind_conversations ("
-    "conversation_id TEXT PRIMARY KEY, "
-    "tags TEXT NOT NULL DEFAULT '{}', "
-    "created_at TEXT NOT NULL"
-    ")"
-)
-
-
-def create_mind_conversations_table(
-    host: OnlineHostInterface,
-    agent_state_dir: Path,
-    settings: ProvisioningSettings,
-) -> None:
-    """Create the mind_conversations table in the llm database.
-
-    Uses CREATE TABLE IF NOT EXISTS so it is safe to call multiple times.
-    The table stores conversation metadata (tags, created_at) that the llm
-    tool's built-in tables do not track.
-    """
-    db_path = _get_llm_db_path(agent_state_dir)
-    cmd = f"sqlite3 {shlex.quote(str(db_path))} {shlex.quote(MIND_CONVERSATIONS_TABLE_SQL)}"
-    with log_span("Creating mind_conversations table in {}", db_path):
-        result = _execute_with_timing(
-            host,
-            cmd,
-            hard_timeout=settings.fs_hard_timeout_seconds,
-            warn_threshold=settings.fs_warn_threshold_seconds,
-            label="create mind_conversations table",
-        )
-        if not result.success:
-            raise RuntimeError(
-                f"Failed to create mind_conversations table: {result.stderr}. "
-                "Conversation features (chat, system notifications) will not work without this table."
-            )
-
-
-def _insert_conversation_record(
-    host: OnlineHostInterface,
-    agent_state_dir: Path,
-    settings: ProvisioningSettings,
-    *,
-    conversation_id: str,
-    tags: dict[str, str] | None = None,
-) -> None:
-    """Insert a conversation record into the mind_conversations table in the llm database.
-
-    The model is not stored here -- it lives in the llm tool's native
-    ``conversations`` table and is set when the conversation is created
-    via ``llm inject -m <model>``.
-    """
-    now = datetime.now(timezone.utc)
-    created_at = now.strftime("%Y-%m-%dT%H:%M:%S.") + f"{now.microsecond * 1000:09d}Z"
-    tags_json = json.dumps(tags or {}, separators=(",", ":"))
-
-    db_path = _get_llm_db_path(agent_state_dir)
-    sql = (
-        f"INSERT OR REPLACE INTO mind_conversations "
-        f"(conversation_id, tags, created_at) "
-        f"VALUES ("
-        f"{_sql_quote(conversation_id)}, "
-        f"{_sql_quote(tags_json)}, "
-        f"{_sql_quote(created_at)}"
-        f")"
-    )
-    cmd = f"sqlite3 {shlex.quote(str(db_path))} {shlex.quote(sql)}"
-    with log_span("Recording conversation in DB for {}", conversation_id):
-        result = _execute_with_timing(
-            host,
-            cmd,
-            hard_timeout=settings.fs_hard_timeout_seconds,
-            warn_threshold=settings.fs_warn_threshold_seconds,
-            label="insert conversation record",
-        )
-        if not result.success:
-            logger.warning("Failed to insert conversation record for {}: {}", conversation_id, result.stderr)
-
-
-def _sql_quote(value: str) -> str:
-    """Quote a string value for use in a SQL statement.
-
-    Escapes single quotes by doubling them, per SQL standard.
-
-    We use manual quoting here instead of parameterized queries because the
-    SQL is passed to the sqlite3 CLI via host.execute_command() over SSH,
-    where parameterized queries are not available.
-    """
-    escaped = value.replace("'", "''")
-    return f"'{escaped}'"
-
-
-def _inject_conversation(
-    host: OnlineHostInterface,
-    settings: ProvisioningSettings,
-    *,
-    model: str,
-    prompt: str,
-    response: str,
-    label: str,
-    llm_user_path: Path | None = None,
-    env_vars: dict[str, str] | None = None,
-) -> str | None:
-    """Run ``llm inject`` to create a new conversation. Returns the conversation ID on success.
-
-    Omits ``--cid`` so that ``llm inject`` creates a new conversation and
-    prints the assigned ID to stdout (e.g. "Injected message into conversation <id>").
-    """
-    env_prefix = f"LLM_USER_PATH={shlex.quote(str(llm_user_path))} " if llm_user_path else ""
-    if env_vars:
-        env_prefix += " ".join(f"{key}={shlex.quote(value)}" for key, value in env_vars.items()) + " "
-    inject_cmd = (
-        f"{env_prefix}llm inject -m {shlex.quote(model)} --prompt {shlex.quote(prompt)} {shlex.quote(response)}"
-    )
-    result = _execute_with_timing(
-        host,
-        inject_cmd,
-        hard_timeout=settings.install_hard_timeout_seconds,
-        warn_threshold=settings.install_warn_threshold_seconds,
-        label=label,
-    )
-    if not result.success:
-        logger.warning("Failed to create {} conversation via llm inject: {}", label, result.stderr)
-        return None
-
-    # Parse conversation ID from output like "Injected message into conversation <id>"
-    stdout = result.stdout.strip()
-    parts = stdout.rsplit(" ", 1)
-    if len(parts) == 2:
-        return parts[1]
-
-    logger.warning("Could not parse conversation ID from llm inject output: {}", stdout)
-    return None
-
-
-def _create_internal_conversation(
-    host: OnlineHostInterface,
-    agent_state_dir: Path,
-    settings: ProvisioningSettings,
-    *,
-    internal_tag: str,
-    display_name: str,
-    prompt: str,
-) -> None:
-    """Create an internal conversation with the matched-responses model.
-
-    Uses ``llm inject`` to create a new conversation, then inserts a record
-    into the ``mind_conversations`` table with
-    ``tags={"internal": "<internal_tag>", "name": "<display_name>"}``.
-    Services can find conversations by querying for the ``internal`` tag value.
-    """
-    llm_data_dir = agent_state_dir / "llm_data"
-    conversation_id = _inject_conversation(
-        host,
-        settings,
-        model="matched-responses",
-        prompt=prompt,
-        response="Confirmed.",
-        label=internal_tag,
-        llm_user_path=llm_data_dir,
-        env_vars=dict(LLM_MATCHED_RESPONSE=""),
-    )
-    if conversation_id is None:
-        return
-
-    _insert_conversation_record(
-        host,
-        agent_state_dir,
-        settings,
-        conversation_id=conversation_id,
-        tags={"internal": internal_tag, "name": display_name},
-    )
-    logger.info("Created {} conversation: conversation_id={}", internal_tag, conversation_id)
-
-
-def create_system_notifications_conversation(
-    host: OnlineHostInterface,
-    agent_state_dir: Path,
-    settings: ProvisioningSettings,
-) -> None:
-    """Create the system_notifications conversation for delivery failure alerts."""
-    _create_internal_conversation(
-        host,
-        agent_state_dir,
-        settings,
-        internal_tag="system_notifications",
-        display_name="System Notifications",
-        prompt="This channel is for system notifications, warnings, and errors.",
-    )
-
-
-def create_slack_notifications_conversation(
-    host: OnlineHostInterface,
-    agent_state_dir: Path,
-    settings: ProvisioningSettings,
-) -> None:
-    """Create the slack_notifications conversation for Slack integration alerts."""
-    _create_internal_conversation(
-        host,
-        agent_state_dir,
-        settings,
-        internal_tag="slack_notifications",
-        display_name="Slack Notifications",
-        prompt="This channel is for Slack notifications and messages.",
-    )
-
-
-def create_daily_conversation(
-    host: OnlineHostInterface,
-    agent_state_dir: Path,
-    settings: ProvisioningSettings,
-    chat_model: str,
-) -> None:
-    """Create a daily conversation tagged with today's date.
-
-    Uses ``llm inject`` to seed the conversation with an empty user prompt
-    and a greeting from the assistant, then inserts a record into the
-    ``mind_conversations`` table with ``tags={"daily": "<today>"}``.
-    """
-    today = datetime.now(timezone.utc).strftime("%Y-%m-%d")
-
-    llm_data_dir = agent_state_dir / "llm_data"
-    conversation_id = _inject_conversation(
-        host,
-        settings,
-        model=chat_model,
-        prompt="",
-        response="Hi, I'm Selene! How can I help?",
-        label="daily",
-        llm_user_path=llm_data_dir,
-    )
-    if conversation_id is None:
-        return
-
-    _insert_conversation_record(
-        host,
-        agent_state_dir,
-        settings,
-        conversation_id=conversation_id,
-        tags={"daily": today, "name": f"Daily Thread ({today})"},
-    )
-    logger.info("Created daily conversation: conversation_id={} date={}", conversation_id, today)
-
-
 def compute_claude_project_dir_name(work_dir_abs: str) -> str:
     """Compute the Claude project directory name from an absolute work_dir path.
 
@@ -738,27 +243,6 @@ def compute_claude_project_dir_name(work_dir_abs: str) -> str:
     absolute path, e.g. /home/user/.minds/my-agent -> -home-user--minds-my-agent
     """
     return work_dir_abs.replace("/", "-").replace(".", "-")
-
-
-def resolve_work_dir_abs(
-    host: OnlineHostInterface,
-    work_dir: Path,
-    settings: ProvisioningSettings,
-) -> str:
-    """Resolve the absolute path of work_dir on the host.
-
-    Returns the absolute path as a string.
-    """
-    abs_result = _execute_with_timing(
-        host,
-        f"cd {shlex.quote(str(work_dir))} && pwd",
-        hard_timeout=settings.fs_hard_timeout_seconds,
-        warn_threshold=settings.fs_warn_threshold_seconds,
-        label="resolve work_dir",
-    )
-    if not abs_result.success:
-        raise RuntimeError(f"Failed to resolve absolute path of {work_dir}: {abs_result.stderr}")
-    return abs_result.stdout.strip()
 
 
 def setup_memory_directory(
@@ -774,29 +258,14 @@ def setup_memory_directory(
     - <work_dir>/<active_role>/memory/ (if it doesn't exist)
     - ~/.claude/projects/<project_name>/memory/ (real directory, not symlink)
     - Initial rsync of contents from role memory/ to claude project memory/
-
-    The project_name is derived from role_dir_abs (the absolute path to the
-    role directory, e.g. ``/home/user/.minds/agent/thinking``), because
-    Claude Code runs from within the role directory.
-
-    Memory sync hooks (added separately via build_memory_sync_hooks_config) keep
-    the two directories in sync during agent operation: PreToolUse syncs from the
-    version-controlled role memory into Claude's project memory, and PostToolUse
-    syncs back so that any memory Claude wrote is captured in version control.
     """
     memory_dir = work_dir / active_role / "memory"
-    # Use .parent because Claude Code's project dir is named after the git repo
-    # root (the mind dir), not the role subdirectory within it. This must
-    # match the path used by build_memory_sync_hooks_config.
     project_dir_name = compute_claude_project_dir_name(str(Path(role_dir_abs).parent))
 
-    # Create both memory directories.
-    # Remove any existing symlink at the project memory path (from old provisioning)
-    # before creating a real directory.
     quoted_project_dir_name = shlex.quote(project_dir_name)
     project_memory_shell = f'"$HOME/.claude/projects/"{quoted_project_dir_name}/memory'
     cmd = f"mkdir -p {shlex.quote(str(memory_dir))} && rm -f {project_memory_shell} && mkdir -p {project_memory_shell}"
-    _execute_with_timing(
+    execute_with_timing(
         host,
         cmd,
         hard_timeout=settings.fs_hard_timeout_seconds,
@@ -804,11 +273,9 @@ def setup_memory_directory(
         label="mkdir memory dirs",
     )
 
-    # Initial sync: copy existing version-controlled memory into Claude's project memory.
-    # Trailing slashes on both paths ensure rsync copies contents, not the directory itself.
     sync_cmd = f"rsync -a --delete {shlex.quote(str(memory_dir))}/ {project_memory_shell}/"
     with log_span("Initial memory sync: {} -> $HOME/.claude/projects/{}/memory", memory_dir, project_dir_name):
-        result = _execute_with_timing(
+        result = execute_with_timing(
             host,
             sync_cmd,
             hard_timeout=settings.fs_hard_timeout_seconds,
@@ -822,27 +289,15 @@ def setup_memory_directory(
 def build_memory_sync_hooks_config(role_dir_abs: str) -> dict[str, Any]:
     """Build Claude hooks config for syncing per-role memory with Claude project memory.
 
-    Takes role_dir_abs, the absolute path to the role directory (e.g.
-    ``/home/user/.minds/agent/thinking``), because Claude Code runs
-    from within the role directory and its project dir name is derived from
-    that path.
-
     Returns a hooks config dict with PreToolUse and PostToolUse entries that
-    rsync the memory directory in the appropriate direction:
-    - PreToolUse: <role_dir>/memory/ -> ~/.claude/projects/<project>/memory/
-      (ensures Claude sees the latest version-controlled memory)
-    - PostToolUse: ~/.claude/projects/<project>/memory/ -> <role_dir>/memory/
-      (captures any memory Claude wrote back into version control)
+    rsync the memory directory in the appropriate direction.
     """
-    # note that the ".parent" is necessary here--the git repo is what is tracked on the claude side
     project_dir_name = compute_claude_project_dir_name(str(Path(role_dir_abs).parent))
     quoted_work_memory = shlex.quote(f"{role_dir_abs}/memory")
     quoted_project_dir_name = shlex.quote(project_dir_name)
     project_memory_shell = f'"$HOME/.claude/projects/"{quoted_project_dir_name}/memory'
 
-    # Pre: sync version-controlled memory INTO Claude's project memory
     pre_cmd = f"rsync -a --delete {quoted_work_memory}/ {project_memory_shell}/"
-    # Post: sync Claude's project memory BACK to version-controlled memory
     post_cmd = f"rsync -a --delete {project_memory_shell}/ {quoted_work_memory}/"
 
     return {
