+++ libs/mng_llm/conftest.py
@@ -1,0 +1,15 @@
+"""Project-level conftest for mng-llm.
+
+When running tests from libs/mng_llm/, this conftest provides the common pytest hooks
+that would otherwise come from the monorepo root conftest.py (which is not discovered
+when pytest runs from a subdirectory).
+
+When running from the monorepo root, the root conftest.py registers the hooks first,
+and this file's register_conftest_hooks() call is a no-op (guarded by a module-level flag).
+"""
+
+from imbue.imbue_common.conftest_hooks import register_conftest_hooks
+from imbue.mng.utils.logging import suppress_warnings
+
+suppress_warnings()
+register_conftest_hooks(globals())
