+++ libs/mng_llm/imbue/mng_llm/conftest.py
@@ -1,0 +1,16 @@
+from collections.abc import Generator
+
+import pytest
+from loguru import logger
+
+from imbue.mng.utils.plugin_testing import register_plugin_test_fixtures
+
+register_plugin_test_fixtures(globals())
+
+
+@pytest.fixture(autouse=True)
+def _reset_loguru() -> Generator[None, None, None]:
+    """Reset loguru handlers before and after each test to prevent handler leakage."""
+    logger.remove()
+    yield
+    logger.remove()
