+++ libs/mng_llm/imbue/mng_llm/resources/ttyd_chat.sh
