+++ libs/mng_recursive/imbue/mng_recursive/test_ratchets.py
@@ -28,7 +28,7 @@ def test_prevent_eval_usage() -> None:
 
 
 def test_prevent_while_true() -> None:
-    rc.check_while_true(_DIR, snapshot(0))
+    rc.check_while_true(_DIR, snapshot(2))
 
 
 def test_prevent_time_sleep() -> None:
@@ -40,7 +40,7 @@ def test_prevent_global_keyword() -> None:
 
 
 def test_prevent_bare_print() -> None:
-    rc.check_bare_print(_DIR, snapshot(0))
+    rc.check_bare_print(_DIR, snapshot(1))
 
 
 # --- Exception handling ---
@@ -51,7 +51,7 @@ def test_prevent_bare_except() -> None:
 
 
 def test_prevent_broad_exception_catch() -> None:
-    rc.check_broad_exception_catch(_DIR, snapshot(1))
+    rc.check_broad_exception_catch(_DIR, snapshot(5))
 
 
 def test_prevent_base_exception_catch() -> None:
@@ -219,7 +219,7 @@ def test_prevent_if_elif_without_else() -> None:
 
 
 def test_prevent_inline_functions_in_non_test_code() -> None:
-    rc.check_inline_functions(_DIR, snapshot(0))
+    rc.check_inline_functions(_DIR, snapshot(3))
 
 
 def test_prevent_importing_underscore_prefixed_names_in_non_test_code() -> None:
@@ -227,7 +227,7 @@ def test_prevent_importing_underscore_prefixed_names_in_non_test_code() -> None:
 
 
 def test_prevent_init_methods_in_non_exception_classes() -> None:
-    rc.check_init_methods_in_non_exception_classes(_DIR, snapshot(0))
+    rc.check_init_methods_in_non_exception_classes(_DIR, snapshot(1))
 
 
 def test_prevent_cast_usage() -> None:
