+++ libs/mng_ttyd/imbue/mng_ttyd/resources/ttyd_agent.sh
